15-10-2015

+++ db/migrate/20150902210135_add_price_to_product.rb
@@ -1,5 +1,5 @@
 class AddPriceToProduct < ActiveRecord::Migration
   def change
-    add_column :products, :price, :float
+    add_column :products, :price, :integer
   end
 end
