+++ db/migrate/20150907090921_create_orders.rb
@@ -3,7 +3,7 @@ class CreateOrders < ActiveRecord::Migration
     create_table :orders do |t|
       t.integer :user_id
       t.integer :product_id
-      t.float :total
+      t.integer :total
     end
     add_index :orders, :user_id
     add_index :orders, :product_id
